"<" didn't work well with struct.pack for some reason. Also fixed #1
Removed "<" form struct.pack because it didn't work. The lookup grid is
improved but not perfect. Each face's bounding box is used to determine
if the face is within a certain cell in the table so faces with diagonal
edges may cover more cells than it needs to. Not a big deal really...